+++ src/AlgoTradeForge.Domain/Trading/Position.cs
@@ -4,10 +4,10 @@ public sealed class Position
 {
     public Asset Asset { get; }
     public decimal Quantity { get; private set; }
-    public decimal AverageEntryPrice { get; private set; }
-    public decimal RealizedPnl { get; private set; }
+    public long AverageEntryPrice { get; private set; }
+    public long RealizedPnl { get; private set; }
 
-    public Position(Asset asset, decimal quantity = 0m, decimal averageEntryPrice = 0m, decimal realizedPnl = 0m)
+    public Position(Asset asset, decimal quantity = 0m, long averageEntryPrice = 0L, long realizedPnl = 0L)
     {
         Asset = asset;
         Quantity = quantity;
@@ -15,8 +15,8 @@ public sealed class Position
         RealizedPnl = realizedPnl;
     }
 
-    public decimal UnrealizedPnl(decimal currentPrice) =>
-        Quantity == 0m ? 0m : (currentPrice - AverageEntryPrice) * Quantity * Asset.Multiplier;
+    public long UnrealizedPnl(long currentPrice) =>
+        Quantity == 0m ? 0L : (long)((currentPrice - AverageEntryPrice) * Quantity * Asset.Multiplier);
 
     internal void Apply(Fill fill)
     {
@@ -33,17 +33,17 @@ public sealed class Position
             if (Math.Abs(newQuantity) > Math.Abs(Quantity))
             {
                 var totalCost = Quantity * AverageEntryPrice + fillQuantity * fill.Price;
-                AverageEntryPrice = totalCost / newQuantity;
+                AverageEntryPrice = (long)(totalCost / newQuantity);
             }
             else
             {
                 var closedQuantity = Quantity - newQuantity;
-                RealizedPnl += closedQuantity * (fill.Price - AverageEntryPrice) * Asset.Multiplier;
+                RealizedPnl += (long)(closedQuantity * (fill.Price - AverageEntryPrice) * Asset.Multiplier);
             }
         }
         else
         {
-            RealizedPnl += Quantity * (fill.Price - AverageEntryPrice) * Asset.Multiplier;
+            RealizedPnl += (long)(Quantity * (fill.Price - AverageEntryPrice) * Asset.Multiplier);
             AverageEntryPrice = fill.Price;
         }
 
@@ -53,7 +53,7 @@ public sealed class Position
     internal void Reset()
     {
         Quantity = 0m;
-        AverageEntryPrice = 0m;
-        RealizedPnl = 0m;
+        AverageEntryPrice = 0L;
+        RealizedPnl = 0L;
     }
 }
